Found problem for failing thread test.
Changed lock to W mode instead of X as linux didn't seem to be able to adopt it.
Added a cache for working dir to fall back on if it's missing.
Changed absolute to use our get_working_dir as Pathlib's absolute method failed the same way.

Fixed leading path delimiter not being removed on posix, made lock work.